FIX add react boundary to catch rendering errors
