+++ package.json
@@ -18,7 +18,8 @@
     "@mui/material": "^5.15.14",
     "@tauri-apps/api": "^1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.13"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^1",
